Create smaller width version of projects page
